Complete requirements for JS3 project
I went through and completed the requirements for this sections project. For the most part, it was not too difficult. There were just a few things with prototypal inheritance that I struggled with initially. My PM helped clarify the missing information.  Once I better understood what was going on I was able to push through and complete all of the stretch goals as well.